No errors

+++ src/lib/phred2prob/Phred2Prob.java
@@ -15,7 +15,7 @@ public final class Phred2Prob {
 	private final double[] phred2baseErrorP;
 
 	// phred capped at 41
-	public static final int MAX_Q = 41 + 1; // some machines give phred score of 60 -> Prob of error: 10^-6 ?!
+	public static final byte MAX_Q = 41 + 1; // some machines give phred score of 60 -> Prob of error: 10^-6 ?!
 	private static Phred2Prob[] singles = new Phred2Prob[BaseCallConfig.BASES.length + 1];
 
 	private Phred2Prob(int n) {
